Do not auto-detect Eclipse when calling the EclipseHelper as main class
There needs to be a way to force the EclipseHelper to index all it can,
even when it thinks it is not running inside Eclipse at all.

The best idea is to make this automatic when calling EclipseHelper as a
main class because we *know* that the developer wants to index the
annotations in this case, even if Eclipse Indigo has passed the class path
elements in a semi-random order.

Wished-for by Christian Dietz.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>