EclipseHelper: handle manifest class path entries correctly
Simply specifying the file name does not make a correct URL, of course...
it needs to be relative to the parent URL.

This was not relevant so far because we only looked at manifest class
path entries of .jar files generated by the surefire-maven-plugin. But
in the next commit, we want to be able to index thoroughly, so we need
to be able to handle bare files instead of full URLs in the manifest's
Class-Path: line.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>