Copy-edit imagej-updater's ByteCodeAnalyzer
This commit brings the ByteCodeAnalyzer from imagej-updater into SciJava
common. The idea is not so much to move it here, but to modify it for
use with the annotation indexer.

To this end, the ByteCodeAnalyzer's fields, methods and inner classes
were all restricted as much as possible, and the class itself was made
package-local, to make clear that this class is not meant to be part of
the official API of SciJava common.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>