Verify that default values are *not* put into the index
This is a subtle issue: if an annotation for a given class does not
specify a particular value, the developer obviously intended the default
value to be used. Now, if the default value ever changes, it should be
reflected in the values returned by the annotation index. That implies
that we must not write out unspecified settings into the serialized
annotations, but always infer them at runtime from the annotation class
itself.

Happily, our annotation processor does The Right Thing (but the
EclipseHelper in its current form does not). Keep it that way.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>