Test that EclipseHelper does not serialize default values
The switch to reading annotations directly from .class files in the
EclipseHelper fixed a subtle bug: default values for annotation
parameters must not be serialized into the annotation index, but instead
be inferred at runtime (in case the default values in the annotation
class itself changed).

Keep this bug fixed by testing for it in the unit tests.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>