EclipseHelper: ignore .jar file that cannot be found
We recently switched to a more thorough indexing when the EclipseHelper
is called as a main class, including inspecting the Class-Path entry of
the manifest of .jar files. However, said entries might not refer to
existing files, in which case we need to ignore them.

While at it, this fixes the bug that .jar files on the class path would
not have been accessed properly if their paths contained spaces.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>