Work around version skews with XML libraries
Sometimes (I am looking at you, OME!), obsolete XML components are
shipped as part of an OMERO-insight installation, intended to support
Java 1.5 which does not have its own XML libraries).

When multiple projects do that (I am looking at you, Batik!), the
XML components can easily cause trouble when the order of the individual
.jar files cannot guarantee that an appropriate xalan/xerces pairings
is discovered via the class path.

Now, in this time and age, Java 6 is thankfully common enough that we
can always fall back to the XML library contained in the JRE. Which
is exactly what this workaround is all about.

Jan Eglinger was very helpful in diagnosing the issue that he reported
in http://fiji.sc/bugzilla/show_bug.cgi?id=749.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>