AbstractModuleItem: mark items required by default
The required attribute of the @Parameter annotation has a default value
of true. The default value of ModuleItem#isRequired() is supposed to
match that. This discrepancy (and hence bug) is probably a holdover
from before all parameters were required by default, long ago.

Without this change, all scripting parameters -- and really *all*
other ModuleItem implementation besides CommandModuleItem -- report
their parameters as required=false by default, which is inconsistent
with commands. It also causes problems with certain forms of
interoperability, such as OMERO, which does not bother to prompt for
parameter values on the command line if they are marked optional.