Correct preprocessor priorities
The LoadInputsPreprocessor needs to run before FilePreprocessor. If it
does not, the FilePreprocessor will pop a file chooser dialog and one of
two things will happen:
* a file path will be selected
* the command will be canceled

In the latter case, the LoadInputsPreprocessor does not run. In the
former, the file path has been resolved already and any persisted value
is irrelevant.

Both of these preprocessors were at Low priority + 1, basically because
they wanted to run before the InputHarvester. Thus their priorities have
been updated to InputHarvester.PRIORITY + 1 (LoadInputs) and + 0.5
(File) to ensure they execute in the correct order.