Help certain callers of the TestUtils class
In certain components, it is very convenient to have utility methods for
testing data structures involving temporary directories. For example,
the updater tests want to create a temporary directory for the
ImageJ.app/ directory, another one for the update site, and then bundle
them into a FilesCollection data structure. Such methods are typically
stored in classes with the suffix ...TestUtils, so let's special-case
those classes and ignore them when trying to figure out the call site.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>