Ensure createTemporaryDirectory() can be called in a loop
When calling said method from the same code location multiple times, we
should still get an empty temporary directory every time, even if there
are still-open files in the directory of the default name. For Windows,
emptying directories poses an insurmountable problem when they contain
files that are still in use, therefore we have to fall back to creating
new temporary directories.

Identified by Jennyver Jenkins as a problem when building and testing
ij1-patcher on Windows.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>