ModuleInfo: add a loadDelegateClass() method
This makes it easier to more consistently and safely obtain the Class<?>
object associated with the meat of the module being described.

Previously, you could call getDelegateClassName() but then there was no
safe way, in general, to load that class in all situations

Previously, you could call getDelegateClassName() but then there was
no safe way, in general, to load that class in all situations.