UIService: Do *not* conflate the 'initialized' and 'disposed' flags
Earlier, this developer tried to be smart (failing) and reuse the
'initialized' flag to indicate whether to show the user interface
or not. However, that flag did not indicate whether the initialize()
method was run yet, but whether the service's lazy initialization was
already performed or not.

Let's just have a separate flag to indicate whether the context was
disposed or not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>