Migrate SCIFIO I/O framework to SciJava Common
This implementation is a restructuring of SCIFIO's io.scif.io package,
to fit within the SciJava plugin framework.

This commit introduces a new plugin type, DataHandle, that defines all
the methods for random access reading and writing. The DataHandle
interface is a rich SciJava plugin type that extends the
java.io.DataInput, java.io.DataOutput and java.io.Closeable interfaces.

Furthermore, the DataHandle interface also defines additional API
methods that make it very straightforward to adapt DataHandle objects
into InputStreams and OutputStreams (it cannot extend InputStream and
OutputStream, because they are abstract classes, not interfaces).

The actual adaptation to streams is done using the DataHandleInputStream
and DataHandleOutputStream wrapper objects, simply by delegating to the
various DataHandle methods of the same name.

Overall, the mapping from the old io.scif.io API is as follows:

* IRandomAccess + IStreamAccess -> org.scijava.io.DataHandle
* RandomAccessInputStream -> org.scijava.io.DataHandleInputStream
* RandomAccessOutputStream -> org.scijava.io.DataHandleOutputStream
* Location -> subsumed into the org.scijava.io.Location type hierarchy
* HandleException -> unnecessary; IOException et. al are good enough
* FileHandle, URLHandle, etc. -> same names, implementing DataHandle

Note that the mapping is rather loose, with several layers consolidated
or eliminated. Additional commits will readd features, such as specific
DataHandle plugins, as appropriate to achieve needed functionality.