Only warn about unmatched inputs in modules
We used to issue an error when the module service's run() method
was called with inputs that are not specified explicitly in the
corresponding ModuleInfo.

That is fragile: should the implementing classes be modified and remove
an input parameter, we should still do our best to continue.

Therefore, let's just warn instead and continue, making the unmatched
input available to the module and just skipping the conversion to
an explicit type (because the info does not specify it).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>