ThreadService: introduce the getParent(Thread) method
Due to the use of a thread pool, it is impossible to determine which
thread called the current one.

Sometimes we need that, though, e.g. when ImageJ 1.x wants to determine
the (thread-local) macro options and runs in a thread that was spawned
through the thread service.

So let's add API to determine the parent thread, so that the ImageJ 1.x
patcher can be extended to support macro options even when the
ThreadService is involved.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>