TestUtils: valid calling code locations do not include .jar files
The whole point of the getCallingCodeLocation() method in the TestUtils
class is to be able to determine a unique location for a directory
inside the target/ directory of the current Maven project.

Therefore, code locations in .jar files are not appropriate: there is
no associated target/ directory.

Thanks, Jaspar Jenkins, for running the regression tests so faithfully
and pointing out that the imagej-plugins-uploader-{ssh,webdav} projects
require this change to run their integration tests properly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>