Record location of tracked objects, too
We will need this later, because we really want to verify, when
compiling usage statistics, which update site (if any) the relevant file
came from. If we cannot deduce the update site, we don't care about the
object's usage, because it is not part of the "known ecosystem" of
modules etc.