Refactor Prefs to PrefService
Using a static utility method for Prefs is limiting - it provides no way
to override preference saving/loading behavior. This creates problems
when there are errors or negative interactions with the Preferences
implementation - for example, in MATLAB there are known bugs regarding
the Java Preferences implementation.

By converting to a PrefService we can now control this behavior more
precisely, as needed. The Prefs utility class will remain for now, but
will delegate to the PrefService when it has been created - allowing the
utility method behavior to be overridden similarly.