DefaultModuleService: include exception in log
When logging the error creating a module, include the exception too.

We have it, so we should pass it on. (If the stack trace is annoying
on the console, the answer is not to eat the exception, but rather
to change on information from the log is emitted to the console.)