ModuleService: add save/load item methods
Added API to save/load ModuleItems. This will allow Contextual classes
to move away from ModuleItem#loadValue and ModuleItem#saveValue, which
are not Contextual and go directly through the Prefs utility class.

Adds a default implementation in DefaultModuleService, using the
PrefService if able (which requires an AbstractModuleItem). If
necessary, still falls back to ModuleItem's saveValue and loadValue
methods.