AbstractPrefService: account for priority
It was discovered that the initialize method of multiple PrefService
impls was being invoked - leading to the LOWEST priority overwriting the
static service field in the Prefs utility class (since it was the last
initialize method to execute).

Added a check for priority so that lower priority services don't
overwrite higher.