ServiceHelper: load EventService more aggressively
Whenever we have @EventHandler annotated methods, we need an
EventService to be loaded as a dependency. Otherwise, even if the
EventService is loaded later, those event handlers will not be properly
registered with it, since at the time registration was attempted, the
EventService was not yet loaded and hence null.

This change obviates the need for DefaultEventService being set to an
absurdly high priority, as explained at:

https://github.com/imagej/imagej/commit/b9abe178b813fb68aa49fdd7515a5e6698436793