EventServiceTest: try to speed up a test
We do not need to _always_ wait 500 ms, but rather can try waiting for
a notification from the other thread. Given the overhead of queuing
things onto the EDT, it turns out that the test thread does end up
waiting before the EDT notifies. So in practice it should cut down on
how long the test takes to run.

Suggested by Johannes Schindelin.