PrefService: remove Preferences imports
The whole point of the PrefService is to allow for any preference
implementation, while the DefaultPrefService uses java.util.Preferences.

Thus the java.util.Preferences imports have been removed from the
PrefService, and the necessary Class and String-based indexing methods
have been added.

Updated the DefaultPrefService implementation and usage of the service
accordingly.