DefaultScriptService: fix service aliases
Previously, if a service depending on the ScriptService called addAlias
during its initialization (which is the case for ImageJ's
LegacyService), it would trigger the ScriptService to initialize its
alias map, including addition of all available services to the map. But
at that point, not all services are available yet. Better is to wait to
add services to the alias map until it is _certain_ that the service
initialization is complete. The ServicesLoadedEvent is perfect for this.