Create ConversionService
Adds a new HandlerService: ConversionService, operating on
ConversionHandler plugins.

The purpose of these handlers is to process a new class,
"ConversionRequest", which signifies the desire to cast or convert an
object of one type to another.

This will take the place of ConversionUtils, as it is an extensible
framework. ConversionHandler plugins can decide what types of conversion
they support and thus override the default behavior (the default plugin
supports primitive : object conversions and cases where the destination
object has a constructor that will accept the source, for example).

Closes https://github.com/scijava/scijava-common/issues/95