ConversionUtils: delegate to ConversionService
The ConversionUtils canConvert and convert methods will now all delegate
to a ConversionService, if available, or a DefaultConversionHandler if
no ConversionService has been set.

The ConversionUtils class will be retained, as it provides some API
which does not need generalization via a Service/plugin framework. All
delegating methods have been marked deprecated however, and consumers
should switch to ConversionService usage if possible.

Updated the AbstractConversionService so that ConversionServices set
themselves as the delegate service in CovnersionUtils on initialization
(only the highest priority will "stick").