WIP: Record usage of all Handler and Wrapper plugins
[This work was originally intended to go hand in hand with the
SJC UsageService. But that service has now been split out of SJC.
As of this writing, it lives in imagej/imagej-usage, although
more properly it should probably go to scijava/scijava-usage.
What is really needed for this to work as desired is for all
PTServices to publish events whenever a handler or wrapper plugin
is handed to a downstream caller -- this constitutes a "usage" of
that plugin, and the imagej-usage/scijava-usage component should
listen for those events and then mark a tick in the right box.]

It is very straightforward when a Handler or Wrapper plugin is being
"used": the service returns the appropriate and specifically requested
Handler or Wrapper instance for the task at hand. So let's count it in
our usage statistics when that happens!

For PTService extensions in general, this makes it easy for subclasses
to register usage of their plugins in whatever way they see fit.