Gateway: remove get(String) method
Obviously, this is a breaking API change. The reason this method is
being removed is that we would like RichPlugin to implement
BasicDetails, which has its own get(String) method with an incompatible
return type. So the Gateway method ultimately has to go.

Upon consideration, the Gateway#get(String) method was not very useful
anyway: what would you do with the (weakly typed) Service once you got
it back? You'd have to cast it to make any calls to its service API
specifically, which would necessitate a compile-time dependency on the
service's artifact, which would obviate the need to call
Gateway#get(String) in the first place -- in that case, you could just
call Gateway#get(Class) instead.

I suppose there is a _rare_ case where you'd want to check for a service
at runtime only, without an explicit compile-time dependency, and then
invoke one or more of its methods via reflection. But this is certainly
not an encouraged practice, so let's not have the Gateway make it easy.
It is still easy enough to do anyway:

    Service s = myGateway.context().service("MyService");