ScriptLanguageIndex: also index on language name
It is intuitive to be able to call
scriptService.getLanguageByName(langName), where langName is what is
given by ScriptLanguage#getLanguageName(). So we should index each
ScriptLanguage not only by its short _engine_ names, but also by its
_language_ name, which in some cases may actually be a different name
than any of those used to refer to the engine itself.

For example, the JRubyScriptEngine has "ruby" and "jruby" as its engine
names, but "Ruby" as its language name.