Add capacity for type aliasing module inputs
Added methods ot the Converter class for identifying the Input and
Output classes that can be converted between, and the populateInputs
method - allowing Converters the potential to point consumers to
potential input objects that can certainly be converted to appropriate
output objects.

Added API to the ConvertService to take advantage of the new Converter methods
- aggregating the potential input objects from converters matching a queried
output type.