CommandInfo: remove now-superfluous code
The PluginInfo base class now implements the Locatable#getLocation() and
Versioned#getVersion() methods in the identical way, so CommandInfo no
longer needs to implement them itself. Note that it _does_ still need to
implement the Identifiable#getIdentifier() method, because the
identifier of commands is different than that of plugins in general.