Deprecate (Class, ...) conversion methods
We would like to move away from these signatures as the process of
conversion implies there is an actual source Object to convert from,
which makes these signatures confusing. They should be unnecessary, and
thus removed in SJC 3.0.

This allows us to focus the functionality of the NullConverter, such
that its canConvert signatures always return true for a null Object, and
false for anything else.

This requires a temporary stop-gap of restoring some null class support
in the DefaultConverter, but only in deprecated methods.