Script Interpreter: keep a record of a 'current command'
The 'current command' is the command that is not yet in the history. It
is very convenient to be able to go back in history, just to have a
look, before continuing to craft the current command (and it would be
annoying if it was lost when going back in history). Of course, if the
user decides to execute a different command from the history instead,
the edits to the 'current command' are lost.

This behavior is most in line with the Unix shell behavior power users
might be used to.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>