Add an ol' classic: ReflectedUniverse, aka CurtisJ
Sure, there is BeanShell. But that would add another dependency to
the project! The ReflectedUniverse is simple, tried and true.

Sure, you almost never want to actually use it, since compile-time
safety is much preferred. But when you _do_ need to do some big
reflection thing, it is quite handy.

For example: suppose you want to remove a dependency from a project,
and deprecate the methods that used it. You can leave the methods as
is, so that they behave exactly the same way, but wrap the offending
code in a ReflectedUniverse to avoid the compile-time dependency. That
way, the deprecated methods will continue to work at least in the case
where the removed dependency is still present on the classpath.