ReflectedUniverse: remove LogService usage
The LogService was only used for debugging, and only to emit the same
information already thrown by the checked ReflectException. In other
words: totally superfluous.

The only really valid usage was in the main method, which is merely a
test driver. That method uses System.out.println elsewhere, so using
System.err.println for that one situation is more consistent anyway.