Change default log level to INFO
The defualt log level was previously set to WARN, meaning INFO-level
logging and above would not be output. This was confusing, as INFO
statements were intended to be used to convey basic, helpful
information.

To that end, the default level has been dropped down to INFO. This also
required updating log.info statements that were outputting DEBUG-style
information to be similarly adjusted to log.debug statements.