Context: try hard not to die in non-strict mode
When Context#inject(Object) is called, there are all kinds of crazy
things that can potentially happen, particularly class loading failures
(e.g., NoClassDefFoundError) due to missing dependencies when scanning
the given object's fields and methods.

When the context is in non-strict mode, we want to try very hard not to
throw an exception, but instead simply log an error, when something goes
wrong during the injection. And we want the injection to proceed
tenaciously, so that if something goes wrong at one stage, other stages
still might complete successfully.

This commit is best viewed with git's "-b" flag.