Add caching method to consolidate scans
The getAnnotatedMethods and getAnnotatedFields methods are highly
repetitive, and the addition of caching exacerbates this code
repetition.

Also, maintaining methods for each object type, and requiring one call
per annotation type is not extensible and results in multiple
superclass/interface hierarchy traversals.

Thus a ClassUtils.cacheAnnotatedObjects method is provided to ensure a
list of Annotation:Object pairs is cached and future requests for these
pairings will result in cache hits.

This consolidates all annotation scraping logic in one method (although
there is still repetition for Fields and Methods right now, it could be
further simplified). It also allows a single traversal of the
Class/Interface hierarchy to discover an arbitrary number of
annotations.