CommandInfo: disallow immutable BOTH parameters
The purpose of BOTH is to indicate than an input reference will be
changed in-place. It is a programming error to label an immutable
parameter as BOTH; rather, the immutable parameter should be an INPUT
and the (presumably affiliated) OUTPUT should be a separate variable.

Arguably, this logic should live somewhere deeper than CommandInfo;
we can cross that bridge later when we come to it.