Put helper classes in the correct enclosing test
The generics methods used to live in ClassUtils, so there were generic
helper classes in ClassUtilsTest to facilitate testing them.

But the methods moved to a dedicated utility class, GenericUtils, and
the unit tests moved along with them to a new GenericUtilsTest.

However, somehow, the helper classes used by those tests did not move
out of ClassUtilsTest and into GenericUtilsTest. Let's do that now.