GenericUtils: add handy getFieldTypeString method
This method gives you the field type declaration which you would need to
use in Java source code as e.g. a method argument, which would be
compile-time type compatible with the field in question in that context.

Many thanks to rgettman on StackOverflow for taking the time to respond
to my question about how to do this:

    http://stackoverflow.com/questions/28143029