Add unit test for validity processing
When a command is invalid, it is supposed to be flagged as such, with
problems recorded. There was a manual test for this in
scijava-plugins-commands called InvalidCommandDemo, which was a silly
place for such a thing. Much, much better to have a unit test here that
verifies this feature is working as expected.

This code was migrated and adapted from:
https://github.com/scijava/scijava-plugins-commands/blob/scijava-plugins-commands-0.1.8/src/main/java/org/scijava/plugins/commands/debug/InvalidCommandDemo.java