Notify listeners of stdout and stderr changes
This change completes the loop, such that stdout and stderr messages
are now reported back to all ConsoleService listeners as desired.

When output occurs, something needs to tell the ConsoleService, so that
it can tell all its listeners what happened. That's where the
OutputStreamReporter comes in: it is added to the MultiOutputStream,
which is wrapped by a MultiPrintStream, which overrides the default
System.out or System.err, such that when output occurs, the
OutputStreamReporter finds out and calls ConsoleService#notifyListeners.

This change also cleans up after itself when the context is disposed,
removing the OutputStreamReporters as appropriate.