Avoid concurrency issues
When streams are added or removed, we do not want that to interfere with
concurrent stream operations (i.e., writing, closing and flushing).
But we also do not particularly want to incur the heavy performance
penalty of synchronizing all the stream operations, either. So let's
keep a copy of the streams, which is regenerated whenever stream is
added or removed from the mix, and use that copy when iterating over the
streams during the stream operations.

Similarly, when listeners are added or removed from the ConsoleService,
we use the same trick to avoid the need to synchronize listener
notification every time an output event occurs.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>