Update concurrency resolution
Instead of nulling out a cached list, we simply replace the cached list
after the core list is updated.

This unfortunately doesn't allow us to be lazy in our caching. But the
lazy caching method was not actually thread safe, because the read of the
cache to check for null state is not synchronized.