DefaultRecentFileService: fix OpenFile package
This will theoretically make the recent file menu entries work again,
though it has not been tested in quite some time. As the FIXME notes,
the current approach is still wrong, in that it requires
scijava-plugins-commands on the classpath for the
DefaultRecentMenuService to function properly. What we should really do
instead is:

* Provide an OpenRecentFile command which leans on the IOService; or
* Migrate the OpenFile command into scijava-common itself.

Either way, not an urgent problem, so the FIXME stays for the moment.