AbstractWrapperService: fix serious wrapping bug
When calling the create method, after a match is found, the data needs
to actually get populated by calling the set method. Otherwise, the
create method will return an unpopulated wrapper object.

This bug was not discovered until now because the only core type of
WrapperPlugin, the InputWidget, overrode the create method with its
own implementation that populated the data properly.