Fix default output type conversion check
As a great man once said, "this behavior is fundamentally wrong."

By default, a converter should match if its known output type is
compatible with the requested output type - not the other way around.

That is, it's OK to return a subclass of the requested type. Returning a
superclass is not guaranteed to be OK.