CommandModuleItem: do not convert empty strings
If the min(), max() or stepSize() parameter values are the empty string,
it means they were unset and should be treated as null/unspecified.

This fixes imagej/imagej-omero#46.

The problem was that String fields, rather than returning null for min
and max, instead were returning the empty string, which was propagated
to OMERO, which really did not like that. Really, the SJC framework
should have been returning null for min/max in that case anyway.