ScriptModule: populate the return value last
This change allows the return value (named "result") to be overridden
as a declared, typed output parameter. Or if left undeclared, it
exists as an implicit output of type java.lang.Object.

This change hopefully prevents bugs surrounding (lack of) conversion
of an output named "result" to its declared type.