ScriptModule: improve handling of return values
Whether the return value output is explicitly declared makes a big
difference. The behavior we want is:

                implicit output          explicit output
           /--------------------------------------------------\
           |   result = 2          |   // @OUTPUT int result  |
 variable  |   1                   |   result = 2             |
 assigned  |                       |   1                      |
           |                       |                          |
           | BEHAVIOR: result is 1 | BEHAVIOR: result is 2    |
           |-----------------------+---------------------------
           |   1                   |   // @OUTPUT int result  |
 variable  |                       |   1                      |
unassigned |                       |                          |
           | BEHAVIOR: result is 1 | BEHAVIOR: result is null |
           \--------------------------------------------------/

The thinking is that when the script writer declares an output called
"result", their intuition will be that the final assignment to the
variable called "result" should become the output value. But when they
_don't_ explicitly declare that output, the implicitly added output
called "result" refers to the return value of the script itself, not
the value of the variable called "result" in the bindings.